+++ src/translations/en/squeeze.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+  'Failed to generate the archive' => 'Failed to generate the archive',
+];