+++ CHANGELOG.md
@@ -1,8 +1,17 @@
-# Release Notes for Squeeze 1.x
+# Changelog
+
+## 2.0.0 - 2024-10-25
+> {note} The plugin’s package name has changed to `verbb/squeeze`. Squeeze will need be updated to 2.0 from a terminal, by running `composer require verbb/squeeze && composer remove olivierbon/craft-squeeze`.
+
+### Changed
+- Migration to `verbb/squeeze`.
+- Now requires Craft 4.0+.
 
 ## 1.1.2 - 2019-05-29
 - Fixed an issue where the temporary zip wasn't being deleted
+
 ## 1.1.0 - 2019-01-17
 - Added ability to trigger download via URL
+
 ## 1.0.0 - 2019-01-17
 - Initial release